Refactorise la page d'accueil et change des chart avec le nouveau composant
