+++ app.territoiresentransitions.react/src/app/pages/collectivite/Accueil/AccueilCard.tsx
@@ -8,7 +8,7 @@ type AccueilCardProps = {
 const AccueilCard = ({children, className}: AccueilCardProps): JSX.Element => {
   return (
     <div
-      className={classNames('h-full bg-white rounded-lg p-8', className)}
+      className={classNames('bg-white rounded-lg p-8', className)}
       style={{
         boxShadow: '0px 2px 16px 0px #0063CB0A, 0px 4px 6px 0px #0063CB0F',
       }}
