+++ app.territoiresentransitions.react/src/app/pages/collectivite/Accueil/AccueilEmptyCardWithPicto.tsx
@@ -14,7 +14,7 @@ const AccueilEmptyCardWithPicto = ({
   picto,
 }: AccueilEmptyCardWithPictoProps): JSX.Element => {
   return (
-    <AccueilCard className="grid md:grid-cols-3 gap-8">
+    <AccueilCard className="grow grid md:grid-cols-3 gap-8">
       <div className="m-auto">{picto}</div>
       <div className="md:col-span-2 flex flex-col justify-end">{children}</div>
     </AccueilCard>
